Refactor: Change name of type file
