+++ components/FeedbackView/FeedbackView.tsx
@@ -7,7 +7,7 @@ import { GiftSuggestion } from '@/app/types/giftSuggestion';
 import { Profile } from '@/app/types/profile';
 import { generateAndUpdateNewGiftSuggestion } from '@/lib/generateAndUpdateNewGiftSuggestion';
 import { useState, JSX } from 'react';
-import { IFeedbackViewProps } from './FeedbackViewProps';
+import { IFeedbackViewProps } from './IFeedbackViewProps';
 
 /**
  * Component that creates feedback view and handles userfeedback. It is used in GiftSuggestionCard.tsx.
